// src/app/model/dashboard.ts
export class Dashboard {
    numberOfDoctor: number;
    numberOfPatient: number;
    totalAmountJan: number;
    numberOfCancelAppointment: number;
    currentListedPrice: number;
    totalNewPatientThisQuarter: number;
    totalNewDoctorThisQuarter: number;
    totalAmountFeb: number;
    totalAmountMar: number;
    totalAmountApr: number;
    totalAmountMay: number;
    totalAmountJune: number;
    totalAmountJuly: number;
    totalAmountAug: number;
    totalAmountSep: number;
    totalAmountOct: number;
    totalAmountNov: number;
    totalAmountDec: number;
    numberOfActivePatient: number;
    numberOfActiveDoctor: number;
    totalAmountThisMonth: number;
    numberOfPaidAppointment: number;
    numberOfUnpaidAppointment:number;
}